Loki: Rename public/app/plugins/datasource/loki/language_utils.ts file to follow convention
