+++ public/app/plugins/datasource/loki/languageUtils.ts
