+++ public/app/plugins/datasource/loki/querybuilder/components/UnwrapParamEditor.tsx
@@ -7,7 +7,7 @@ import { Select } from '@grafana/ui';
 import { getOperationParamId } from '../../../prometheus/querybuilder/shared/operationUtils';
 import { QueryBuilderOperationParamEditorProps } from '../../../prometheus/querybuilder/shared/types';
 import { LokiDatasource } from '../../datasource';
-import { isBytesString } from '../../language_utils';
+import { isBytesString } from '../../languageUtils';
 import { getLogQueryFromMetricsQuery, isValidQuery } from '../../query_utils';
 import { lokiQueryModeller } from '../LokiQueryModeller';
 import { LokiVisualQuery } from '../types';
